update notebook; add test
