+++ src/tests/unittests/algorithms/test_functions.py
@@ -1,13 +1,21 @@
-""" Mathematical functions and models."""
-
 import unittest
 import matplotlib.pyplot as plt
 import numpy as np
+import qtt
 from qtt.algorithms.functions import gaussian, fit_gaussian, fit_double_gaussian, double_gaussian, exp_function, \
     fit_gauss_ramsey, gauss_ramsey, cost_exp_decay, logistic, linear_function, Fermi, fit_exp_decay, _estimate_exp_decay_initial_parameters
 
 
 class TestFunctions(unittest.TestCase):
+
+    def test_estimate_dominant_frequency(self):
+        y_data = np.array([0.122, 0.2  , 0.308, 0.474, 0.534, 0.618, 0.564, 0.436, 0.318,
+               0.158, 0.13 , 0.158, 0.336, 0.434, 0.51 , 0.59 , 0.592, 0.418,
+               0.286, 0.164, 0.156, 0.186, 0.25 , 0.362, 0.524])
+        sample_rate = 47368421
+        ff=qtt.algorithms.functions.estimate_dominant_frequency(y_data, sample_rate = sample_rate, fig=1)
+        self.assertAlmostEqual(ff, 5684210, error=10)
+        plt.close(1)
 
     @staticmethod
     def test_fit_exp_decay():
